watcher/requirements.txt
google-cloud
google-cloud-storage
functions-framework==3.2.0
pyaes
pbkdf2
pycryptodome